requirements.txt
black==23.12.0
isort==5.13.2
nox==2022.8.7
pytest==7.4.1
pandas==2.0.3
pandas-stubs
mypy==1.5.1
PyYAML==6.0.1